Move svg org validation to schema
It's better to catch bad input ASAP. Originally the svg and name
validation for an organization was put in the model because there
was no higher level place where orgs were created. Now orgs are
created in the admin panel so it makes sense to move this validation
to the admin panel. It's also generally best to keep as much logic
as possible out of the model so again it makes sense to move it.

Failure messages were also improved to be more explicit about why
exactly the svg was invalid.

Tests were also updated to be more explicit about checking for not
just the colander.Invalid but also the message.