Add a partial index on the `user.nipsa` column
Add an index for efficient lookup of shadowbanned users. This is
currently used during reindexing operations and on the NIPSA admin page.

Since we are only interested in optimizing lookup of users who _are_
shadowbanned, add a predicate on the index to keep it small.

The DB query in `NipsaService.fetch_flagged_userids` has been modified
slightly to make sure it uses exactly the same predicate that is used
when building the index. Otherwise the index doesn't get used.

See https://www.postgresql.org/docs/9.6/static/indexes-partial.html and
http://docs.sqlalchemy.org/en/latest/dialects/postgresql.html#partial-indexes

See https://stackoverflow.com/questions/9822154/ for difference between
`is` and `=` in SQL.