Ignore `ELASTICSEARCH_CLIENT_TIMEOUT` env var in `hypothesis init` command
The request timeout is typically set to a short value (<1s) in
production environments to prevent long-running search/indexing
commands.

The process of initializing the search index can take longer than this
though, so override the default timeout using the same method that is
used for the `hypothesis search reindex` command.