Default to es6 lib in Client class & only test es6
Default to elasticsearch library as opposed to elasticsearch1 in
the Client class.

No longer test elasticsearch1 in the Client tests. This also means
removing the aws tests as we do not use aws auth in elasticsearch6.

Remove es version specific tests and break out the property testing
into seperate tests.