Fix invalid references in docs
Since our source code is not extracted into our Sphinx docs,
cross-references like:

    :py:mod:`h.models`

etc will not work. Sphinx doesn't know what `h.models` is. Sphinx will
style it as a cross-reference in the HTML & CSS, but it won't be able to
hyperlink it, and it'll output a build warning.

These kinds of cross-references *are* valid in our docstrings, though.
Just not in the `docs/**/*.rst` files.