Add migration to set organization NULL for private groups
In the past, we had associated private groups (those groups with
`ReadableBy.members`, `JoinableBy.authority`)
with the default organization (organization with
`pubid` = `__default__`) because the
group -> organization relationship was non-nullable.

Now that a group may have a null organization, set all private group
organizations to `NULL`, as they were put "into" the default organization
simply to satisfy the DB constraint of having an organization.