Install tox when running docs-related make commands
Automatically install `tox` in the developer's virtualenv when running
any of the various docs-related `Makefile` commands.

This fixes an issue that, if the developer didn't already have `tox`
available on their system, the various docs-related `Makefile` commands
will fail.

Having the `Makefile` explicitly install `tox` is consistent with how
`make test` already does it.

The use of `PIP_REQUIRE_VIRTUALENV` in the `Makefile` prevents `tox`
from being installed if the developer hasn't created and activated a
virtualenv for h - it won't try to install `tox` globally.