Add search_after in search query & tests
Previously the param "offset" was being used to set the
"from" property in the search query to search from "offset" annotations to
"limit" number of annotations. This will be deprecated.

Going forward, it will use the param "search_after" to set the
"search_after" property in the search query to search starting at
sort field "search_after" to "limit" number of annotations.

search_after does not run through the parser to format the dates so we
must run the datetime parsing before it is passed to search_after. search_after
only accepts the number epoch_miliseconds as that is the raw format in es for
the created and updated fields.

Add tests for search_after.