Remove the majority of the old-style search query tests
Remove the tests in old_query_test.py which cover functionality that is
also covered by the tests in query_test.py.

Add a note about one aspect of searching by URI which is not explicitly
covered in query_test.py and a note about needing to move the remaining
`Builder` tests.