Record which /api/search params are used
Add custom metrics to record if the offset param is being used-the
idea being we will know if people are still using it as it will be
deprecated and perhaps when no one is using it anymore we can safely
delete it.

Add custom metrics to record the use of tag and tags. tags is an alias
of tag. The idea being if no one is using it we could safely remove this
and simplify our search query params.

Add custom metrics to record the use of url and uri. url is an alias
of uri. The idea being if no one is using it we could safely remove this
and simplify our search query params.