Validate search query params in /api/search
Call validate_query_params in /api/search.

Change the type of request.params used in the tests from a dict
to a NestedMultiDict since this is what type request.params actually
is.

Since the SearchParamsSchema specifies defaults for certain params,
test_it_searches has to be changed to assert the default params were
passed into search as opposed to an empty dict.