Ignore/strip offset when search_after is specified
Offset has been deprecated in favor of search_after. This means both
the Limiter and the Sorter have to handle the case where both offset
and search_after have been specified. This is because we make no
guarantee that the Sorter will be run before the Limiter or vise
versa.

If the Sorter is run first the offset should be popped.

If the Limiter is run first the offset should be popped and ignored.

Add tests for both cases ^.

Cleanup the extract offset and limit logic.