Fix RSS stream
The `/stream.rss` and `/stream.atom` views were broken by
6c342c7cbffaa6efafd883df37237886d2ead706 because the `params.copy()`
call in `h.search.query.Builder.build` was removed. As a result
`Search.run` tried to modify its input `params` dict which crashed in
views were that argument is the read-only `request.params`.

It seems less surprising for `Search.run` not to mutate its input dict,
so restore that behavior.