Validate search_after interactions in the schema
Validate at the schema level the following:
 - Validate that search_after is the correct format based on the value
   of sort.
 - Default offset to 0 if search_after is set.

Add schema tests to validate the above additions to the schema validation.

Remove the handling of both offset and search_after being set in the
Sorter and Limiter classes since this is now handled by the schema.

Remove query tests that validate the handling of both offset and
search_after being set since this is handled by the schema now.