Remove code that establishes elasticsearch_dsl's default connection
We are using elasticsearch_dsl for generating and executing search
queries in app code and tests, but we've opted to stick with the
existing method of establishing a connection to Elasticsearch rather
than using elasticsearch_dsl's default connection.

The existing method of creating a connection and exposing it via
`request.es` is convenient to work with in tests, since it can be easily
mocked. It is also more consistent with how we set up and expose
connections to other external services such as statsd or the database
session.