Split json_error exception view into API and non-API request handlers
We have API and non-API views that return JSON. For consistency handle
unexpected exceptions (aka. Internal Server Errors) for API views in
a view defined in `h/views/api/exceptions.py` and for other views in a
view defined in `h/views/exceptions.py`.

Right now the only difference is that the API view adds CORS headers to
the response.

I have also changed the error message to make it shorter and thus be
easier to read if it gets displayed to the user directly, eg. as the
Hypothesis client will do in a toaster at the top of the sidebar.