Add check_icu_plugin param to search.init
Instead of passing in a full dictionary of settings into search.init,
change search.init to accept a single boolean value for check_icu_plugin
and default it to True. This way the settings don't get passed around the
code, it is backwards compatible, and is a lot less invasive of a change.