Allow "uri.parts" as a query param argument to /api/search
This existing query param was missed when validation of query params to
the search API was added recently.

Many users of this parameter will be better served by the upcoming
wildcard URI search support, but since there were users of the existing
query param, it makes sense to continue to support it.