Fix an SQLAlchemy unicode warning
We're reading the value of the X-Forwarded-User header, which in Python
2 is a byte string, and then using it in an SQLAlchemy query which is
triggering a unicode warning from SQLAlchemy:

    Unicode type received non-unicode bind param value

This is because we're passing SQLAlchemy a byte string and asking it to
match it against a unicode column.

Fix this by decoding the byte string to unicode first.

Once we've moved to Python 3 we can remove this decoding: request header
values are unicode already in Python 3.