Use popall in es search query modifiers (#5304)
* Use _pop_param_values in es search query modifiers

Previously each modifier had essentially the same code to extact
all values of the same key from the params multidict, now they just
call _pop_param_values(key) instead.

Some of the tests were passing in a dict-type value for params instead of
a MultiDict-type. This change forces the params to now be a MultiDict as it
uses a MultiDict-only method called getall().

* Set params type MultiDict in Search.run doc-string

* Rename _pop_param_values to popall

* Change Search to no longer make a copy of params

Search previously made a copy of the params that were passed to it.
This functioned mainly as a way of casting a NestedMultiDict type (readonly)
coming from request into a MultiDict type (mutable). Search calls
query filters that pop the params. In half of the cases where Search
is used, the params that are passed in are already in a poppable state
and copying is expensive thus let the code that calls Search decide whether
to copy/cast the params or not.

The badge view and tests were updated to pass a MultiDict.

The feeds view was updated to cast the request.params into a MultiDict.

The test for verifying the Search doesn't modify params was removed as this
is now expected behavior.

Update the description of params to explain that they will be popped.