Use elasticsearch to get group_annotation_count
Previously the count was obtained from a db query. The problem with
this was two fold:
 1. Shared was being queried on even though it wasn't indexed
 2. When searching for groups such as Public, where the group has a
    large portion of the annotation in the table, the db was optomizing
    to search the table directly rather than use the index due to the
    overhead of indexing lookup.

Since the count is now coming from elasticsearch with the default query
filters, this count now takes into account other things such as; nipsa,
permissions, moderation, etc. While the count previously considered
"annotations" to include replies, since this is used on activity pages
only, it was agreed by the team that it makes more sense for this to
not include replies and exactly match the "Matching Annotations" count.
This decision of not including replies also has the advantage that if
the query run in the search boxc doesn't include any other search
criteria we can re-use that total and we don't have to run this query
at all.