Make subscriber failures independent of one another
When using `request.registry.notify`, an exception in one event
subscriber can cause other subscribers not to run. The order in which
subscribers are run is not specified. This makes it hard to reason about
the eventual state if any one subscriber fails to run.

This commit replaces `request.registry.notify` in
`request.notify_after_commit` with equivalent code using lower-level
zope functions, which wraps each subscriber in a catch block. Therefore
one subscriber failure won't affect others.

Note that there are other uses of `request.registry.notify` in the code
which are not affected by this change. In those places an exception will
cause a database transaction rollback; behavior which we don't want to
change.

The tests were changed to actually dispatch events rather than mocking
`request.registry.notify`. Since the code is now more dependent on
low-level APIs, this will catch any breaking changes in future Pyramid
releases.