Use total ann from search for group total if able
If the search query is only for group and doesn't apply any other
filters, then the total of that search will be the same as the total
returned by the annotations_stats_service so skip the unnecessary
second query by not calling the stats service.

Cleanup mocked annotation_stats service in activity tests to use patch
and autospec.

Add test to verify annotation_stats service isn't called when the total
annotations can be re-used from the search query.