Make AnyMatcher use default AND operator
Previously, the default_operator was left at the default which was OR
so all search terms were ORed together. This meant that as more terms
were added the search broadened in scope/results rather than the intuitive
behavior of narrowing. This has been fixed by defaulting the operator to
AND.

Fixes https://github.com/hypothesis/h/issues/5338