Use elasticsearch to get annotation_user_count
annotation_user_counts previously returned the total number of annotations
a user made, public annotations, and private annotations. Private
annotations was never used and public vs total was used based on whether
the logged in user was the user who had made the annotations. This behavior
essentially re-creates the default behavior of search (return total number
of annotations that the user made based on the permissions of the logged in
user). Thus, when switching from using the db to es, the logic to return
different counts based on the logged in user and the returning different
counts in anootation_user_counts was no longer needed and removed.

Rename annotation_user_counts to annotation_user_count. Since this
method is no longer returning multiple counts but just one, we can
drop the s from the end. This also aligns more closely with the other
method name in this class; annotation_group_count.

Cache the parsed query params from query.extract so that the parser
doesn't have to be run twice in the group and user controllers.
Previously, the params were extracted in SearchController.search and not
returned so the group controller had to recreate them with another call
to query.extract. Now they are created in the constructor and re-used
throughout the class.

Only call annotation_stats.annotation_user_counts if the search
query that was already run contains other search terms (aka: the
total annotations returned from search is different than what would
be returned from annotation_stats). This will save time by not
running the same query twice if it's unnecessary.

Since the query params are now cached in controller.parsed_query_params
some of the activity tests had to be adjusted to mock this value directly.
This is because the query fixture is a dependency of the controller fixture
so it must be mocked with the desired extract.return_value before the controller
fixture is created (in some cases where query.extract.return_value does not
contain the desired default, the default must be overriden directly via
controller.parsed_query_params after the controller is instantiated by the
controller fixture).