Remove tox-pip-extensions
Revert three commits in order to remove tox-pip-extensions from the
project:

1. Revert "Pin to tox 3.3.0"
   This reverts commit dd9aa5302220088d492397b8bf9266cb6c2c8062.

2. Revert "tox: Crash if tox-pip-extensions not installed"
   This reverts commit acff0bc6220fa19fa098276260a4c8a9200e57fe.

3. Revert "Use venv_update to auto update virtualenvs"
   This reverts commit ebaf18690e2d24cd1db466caf8d362638abf2ea6.

We've been having a couple of different tox problems that seem to be
caused, directly or indirectly, by tox-pip-extensions. Removing it for
now to get the build working again. May add tox-pip-extensions back
later when we can get it working, or replace it with something else.