Fix `hypothesis` commands in docs
Fix various `bin/hypothesis` commands in the docs. Now that we use a
tox-managed virtualenv instead of a manually created and activated one
(see https://github.com/hypothesis/h/pull/5261) it's necessary for
`bin/hypothesis` commands to be run inside tox.

Specifically, this commit changes the docs to run the hypothesis command
inside the `py27-dev` virtualenv, which is the virtualenv that
`make dev` uses.

Running the commands like this
(`tox -e py27-dev -- sh bin/hypothesis ...`) is pretty awkward. I'm open
to ideas for how to make it more convenient. For now, though, this gets
the commands in the docs working.