Add a copy icon to annotation url on activity page
The "pretty url" displayed under the URL heading of an annotation
card on the activity page is confusing to users since it is not
the full url. If the user copies this and pastes it into the url
search facet they expect it to work but in fact it does not because
it is not the full url.

Re-use the copy UI for group join links for the annotation URL so
the user can copy the full link inside the input box or simply
click the copy button to copy the full link to the clipboard.