Update bootstrap to fix a vulnerability reported by GitHub
Bootstrap v4 also requires popper.js as a peer dependency.

Bootstrap is only used on the admin pages and then only in a relatively
lite way. I haven't closely checked every usage but have done some basic
manual testing of the navigation on the admin pages and the various
forms.