fixup: remove certifi from requirements.ini
certifi used to be a dependency of h and was used to create the connection
to amazon's es cluster, however that was removed when we upgraded es.
While certifi is a dependency of requests it is not a dependency of h
and thus should be removed from the .ini file.