Update logging filters to reflect Elasticsearch exception changes
When h moved from Elasticsearch 1 to Elasticsearch 6, the internal
library that is used for making requests to the cluster changed from
requests to urllib3. As a result the exceptions that get raised in the
event of network timeouts are from the urllib3.exceptions module instead
of requests.exceptions. See eg. https://sentry.io/hypothesis/h/issues/710947234

This also removes an unnecessary dependency on the "requests" library
inside the tests.