Pass correct exception object to Sentry in exception view
Exception views call `h.util.view.handle_exception` to report the exception to
Sentry. This function calls `request.sentry.captureException()` without
specifying an argument, which means that the Sentry client defaults to capturing
the _last_ exception that was raised. In rare circumstances this is not the
exception that caused the request to fail. Pyramid internally raises, and
catches, exceptions (eg. `PredicateMismatch`) during the routine course of
resolving and invoking the exception view.

The scenario where the `PredicateMismatch` exception is reported to Sentry
instead of the real exception happens only sometimes when running h.
Perhaps there may be non-determinism in the order in which Pyramid
matches requests against candidate view callables?

Resolve the problem by always passing the exception view's context object to
`request.sentry.captureException`. The context object is the exception which
actually caused the request to fail.