Fix Sentry reporting in `h.util.view.handle_exception`
I didn't read the `raven.captureException()` documentation properly. The
first argument is an `exc_info` 3-tuple, not an exception object.

If the exception passed to `handle_exception` is the most recent
exception in the current thread, we can just call `captureException`
without arguments. If it is an older exception, we must construct the
`exc_info` tuple manually. In that case the traceback is only available
in Python 3 via the `exc.__traceback__` property.