Append UriCombinedWildcardFilter in Search class
Previously, because the original UriWildcardFilter was not combined,
the WildcardFilter and the UriFilter were moved outside of the Search
class. Since the UriCombinedWildcardFilter can act as both the original
UriFilter and the new WildcardFilter, move it back into the Search __init__
and add a new param separate_wildcard_uri_keys to Search __init__ that
will be passed to the UriCombinedWildcardFilter. This param will configure
the UriCombinedWildcardFilter to behave in one of two ways:
 - Expect wildcards and exact matches in url/uri keys.
 - Expect wildcards in wildcard_uri and exact matches in url/uri keys.

The default behavior is to separate wildcards and exact uri matches into
separate keys.

Since the UriCombinedWildcardFilter is now appended in the Search class,
places that were previously appending UriFilter or UriCombinedWildcardFilter,
no longer need to do so.

The only view that expects wildcards and exact uri's to be in the same key
is activity pages and a comment was added to explain why separate_wildcard_uri_keys
is set to False.

Add tests for the new param sepatate_wildcard_uri_keys in the Search class.