Add IPython to the dev env
Add IPython and ipdb to the development virtualenv. This means that
`make shell` runs an ipython shell instead of a standard Python shell.

ipdb means you can do `import ipdb; ipdb.set_trace()` in a source code
file and when it hits that line that dev server will stop at an IPython
shell instead of a standard Python shell for debugging.

IPython is a much nicer and friendlier shell than the standard Python
shell.