Update normalize.css
This fixes several cross-browser inconsistencies which lead to visual
bugs in less well tested browsers, such as IE.

Replace previous partial copy of normalize.css with the complete version
of the style from npm. In order to consume styles from npm packages, it
was necessary to add node_modules/ to the style search path.

Importing a plain .css file unfortunately triggers a deprecation warning in the
current version of node-sass [1]. The official solution for that is to
switch to the Dart version.

[1] https://github.com/sass/node-sass/issues/2362

Co-authored-by: Robert Knight <robertknight@gmail.com>