Add rate limiting per user & POST /api/annotations
Rate limit endpoints that are known to cause issues if over-requested
on a per authorization token basis. Use the authorization token instead
of ip address since users from a university may have the same ip address
and the bulk of users in h day-to-day are students.

The main endpoint that is known to cause issues is /api/annotations:create.
When over requested it can overstress the db with create requests causing
long query times which ultimately hog up the gunicorn worker time which
would otherwise be <1%.