Add custom 429 response & multiple zones
Add custom 429 json response for api requests.

Add multiple zones so that quotas on one request don't
impact another request. Re-using a zone means the queue's are
shared and that's not what we want so make different zones
for each endpoint.

Re-order the rate limits so that it follows and if, else format
so it's easier to read.