WIP - Populate search result buckets lazily on activity pages
Speed up activity pages by only rendering the bucket headings in the initial
render. Bucket contents are fetched and rendered when the user clicks, or
begins to click, on a particular bucket.

Activity page rendering time is dominated by template rendering.  Rendering
annotation cards lazily avoids the overhead of generating, transmitting and, on
the client side, parsing large amounts of markup which will not be seen on many
visits.

On fast connections and when bucket rendering is quick there is no visible
difference to the user. On slower connections or if rendering a bucket is slow
on the server, the user will briefly see a loading state after clicking on a
bucket.

This change is feature-flagged.

TODO - Additional search bucket controller tests
     - "activity.bucket" view tests
     - ARIA status updates while loading
     - Consider how to handle JS load failures and non-JS visitors
       (eg. the non-JS Googlebot)
     - Gather some initial payload metrics with and without the feature
       flag enabled