Copy only the files needed for the node build
COPY . . causes a copy of everything into the docker build
which means if any of the files change then that build step
will be re-run. Instead, only copy the things needed for that
build step which in this case are the node package files,
h/static, and the gulp files.