Run gunicorn from newrelic run program
Previously, dev was different than production in that dev was
not run from newrelic so newrelic did not work in the dev environment
and required modifying these two lines manually. Now, with this
change a developer can test with newrelic without modifying code and
simply setting the environment variables in their dev env.