Cleanup metrics (#5473)
* Use one function to record search query params

Previously there were two functions to record search query params,
one of them was for activity pages, the other for the api. The activity
pages use a subset of the params used by the api thus the api function
contains the superset. Instead of having two separate functions called
in each view, have one function that is called in
h/search/core.py:Search.run.

Also instead of wrapping the newrelic add_custom_parameters, call it
directly in the function where it is used and mock newrelic.agent in
the tests. This will simplify the code and reduce abstraction improving
readability.

NewRelic provides an add_custom_parameters function that accepts multiple
parameters at once. The advantage of using this is internally get current
transaction won't be called each time we add a new parameter which will be
a slight performance improvement over calling add_custom_parameter multiple
times.

* Add check for if current transaction exists

Occasionally on startup there is a race condition where the current transaction
may not exist and current_transaction may return None.
If the current transaction is None then there is no add_custom_parameters
function to call and parameters should not be added.

* Make separate_replies key a separate input param

Since separate_replies is actualy a config setting rather than part
of the search query params inside the Search class, pass it in as a
separate argument to avoid adding logic into the Search class. The
logic required to add it as part of the params would modify the params
multidict which is a little dangerous since if done incorrectly could
lead to a different search query.

* Reformat tests

* Minor tweaks based on code review feedback

- Cleanup docstrings
- Break up test
- Rename tests
- Cleanup comments