Add a separate build stage for Python dependencies
The Docker build currently has to go through some hoops to install
packages (eg. C compiler) that are needed to build the Python packages,
but then remove them so they are not included in the final Docker image
or any of its layers.

Recent versions of Docker provide a nicer way to do this using
multi-stage builds. We've already used this for the node build. This
commit applies the same idea to the Python build.

A side benefit is that the final Docker image shrinks slightly from
~158MB to ~141MB, and Docker builds are a little bit faster when Python
dependencies change, as only the pip install has to be repeated, not the
installation of apt packages.

Note that "setuptools" and "git" are still required as runtime
dependencies, even though intuitively they would appear to be only
build dependencies.