Remove git from Docker image
The Git version is written into `h/_version.py` as part of the Docker
build process by `git archive`.

This required a change to `h/_version.py` to handle the case where
running `git` fails because the command could not be found, as well as
the already-handled case where it returns a non-zero status.