Fix flakey test_it_limits_number_of_annotations
The test_it_limits_number_of_annotations had a flakey failure due to
the minimal time between time.now() calls. The population of the
updated field happens so fast that occasionally, the dates are the
same and elasticsearch puts them in a different order. Since this
test depends on the updated fields to be in a certain order (since
updated is the default sort setting), it's safer to populate that field
in the test such that no two annotations have the same updated
value.