Remove unnecessary fixture dependencies
As pointed out in code review, indexing / batch indexing only requires
the moderation service to be mocked, not the group service.

The `Search` class for executing searches only requires the group
service to be mocked, not the moderation service. Many tests involving
search will also do indexing, but the indexing functions will pull in
the moderation service mock if necessary.