Refresh when deleting annotations
When deleting an annotation with delete_by_query, the deleted annotation
is logged in the translog but this change doesn't get pushed out to
the segments and get returned by search until a refresh occurs. By default,
a refresh happens each second which is fine for human time but for machine
test time, the refresh does not happen fast enough.

A version conflict can occur when a document has been updated but
the update has not been propogated to all shards; meaning the current
shard has a lower version number than a different shard that updated
the document. The comment in the code was updated to explain the
reason why this error was happening.