Fix inconsistency in moderation status read/write permissions
When annotations are rendered to JSON for API responses, moderation
information such as the count of flags was included if the user had
"admin" permissions on the annotation's group.

However the API views for performing moderation-related actions tested
for the "moderate" permission on the annotation, which is inherited from
the "moderate" permission for the group.

User accounts assigned to the Staff or Admin roles have the "admin"
permission on all groups but not the "moderate" permission. As a result, staff
accounts on hypothes.is could see the flagged status for annotations in
groups they did not create, but attempting to click the "Hide" button in
the client to moderate the annotation would fail.

The fact that staff accounts have "admin" permissions on all groups but
not "moderate" permissions is an inconsistency in itself, but separate
to the inconsistency in which permission is used when reading/writing
moderation info for annotations.