Fix Celery worker logging on QA and production
Fix sending log messages from Celery tasks to the logs in QA and
production. These weren't being logged at all.

You have to pass a log level to the `hypothesis celery worker` command,
as we already do in dev:

https://github.com/hypothesis/h/blob/4e45a8f18f8a1a002ae7d28458f89dc793880a2e/h/cli/commands/devserver.py#L100

otherwise it doesn't log anything.

Also move the log message when sending an email fails due to an invalid
email address back down to INFO level, where it belongs. (I had moved it
up to WARNING in an attempt to get the message to show up in the logs.)