Fix account registration flow
The account registration flow has been broken since the addition of
the wordpress front page (a couple years). This fixes the broken
registration successful flow.

It also catches a db Integrity error and raises it as a ConflictError in
the signup service. Integrity errors occurr when two signup requests
for the same account race eachother to the database. This is generaly
caused by the user clicking the submit button on the signup page multiple
times. In this scenario the user is informed that they already have an
account and to check their email to confirm it (simillar to the successful
registration flow).