Mark supervisor as a runtime dependency of h
Even though this dependency is only used when running h from the Docker
container, adding it to the declarative list of Python dependencies in
requirements.txt allows Dependabot, GitHub security alerts etc. to be
aware of the dependency. It also means that we can ensure the version
used when building the Docker container does not unexpectedly change if
a new version is released.

As part of this, set the minimum requirement to the recently released v4.0.0
which brings support for Python 3.