Remove @pytest.mark.functional
This is an unregistered custom pytest mark, which causes pytest to omit
a warning and fail the tests. There doesn't seem to be any reason to
try to manually apply this mark to every single functional test -- these
tests are in a separate directory anyway. So rather than registering the
mark, delete it.