Whitelist Hypothesis-Client-Version header in CORS requests
Add "Hypothesis-Client-Version" to the list of headers that browser-based h
API clients are allowed to send in requests.

This header is set by recent versions of the Hypothesis client and is
needed to enable hosting custom builds of the client on domains other
than the one that h is hosted on.