Preserve SVG viewBox when minifying SVGs for production
By default svgo removes the `viewBox` attribute of the SVG, which breaks
scaling of the SVG content when rendering.

This caused various icons on the site to render at an incorrect scale.
In some cases this was not very noticeable because the natural size of
the content was close to the scaled size. In other cases the icon failed
to appear at all.

To see the difference, compare:

```
rm build/images/icons/*
NODE_ENV=production gulp build-images
cat build/images/icons/groups.svg
```

Before and after this patch is applied.

Fixes #5656